requirements.txt
dj-database-url
dj-static
Django
django-cors-headers
django-extensions
django-toolbelt
djangorestframework
djangorestframework-csv
enum34
gunicorn
pandas
psycopg2
python-dateutil
requests[security]
Scrapy
scrapy-djangoitem
pyOpenSSL
